feat: expose clientside module
